API-7793 - Reorganize ActiveModel Validation logic
