+++ modules/claims_api/lib/claims_api/v2/params_validation/veteran_identifier.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+require 'claims_api/v2/params_validation/veteran_identifier/main'
+
+module ClaimsApi
+  module V2
+    module ParamsValidation
+      module VeteranIdentifier
+        def self.validator(params)
+          VeteranIdentifier::Main.new(params)
+        end
+      end
+    end
+  end
+end
